pyproject.toml
[project]
name = "stresstestercli"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.136.1",
    "httpx>=0.28.1",
    "openai>=2.32.0",
    "pydantic>=2.13.3",
    "python-dotenv>=1.2.2",
    "rich>=15.0.0",
    "typer>=0.25.0",
    "uvicorn[standard]>=0.46.0",
]

[build-system]
requires = ["hatchling>=1.26"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
pythonpath = ["src"]

[project.scripts] 
review = "stresstestercli.cli:main"
